requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
isodate==0.7.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
